utils: Move group_by_all from signalmanager to utils.
